consolidate shared UI into @ui components, clean up layout and remove duplication
